% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{get_conditions}
\alias{get_conditions}
\title{Get rule conditions}
\usage{
get_conditions(object, penalty.par.val = "lambda.1se")
}
\arguments{
\item{object}{object of class pre}

\item{penalty.par.val}{character. Value of the penalty parameter value 
\eqn{\lambda} to be used for selecting the final ensemble. The ensemble 
with penalty parameter criterion yielding minimum cv error 
(\code{"lambda.min"}) is taken, by default. Alternatively, the penalty 
parameter yielding error within 1 standard error of minimum cv error 
("\code{lambda.1se}"), or a numeric value may be specified, corresponding 
to one of the values of lambda in the sequence used by glmnet,
for which estimated cv error can be inspected by running \code{x$glmnet.fit}
and \code{plot(x$glmnet.fit)}.}
}
\description{
\code{get_conditions} returns rule conditions in a matrix form
}
\examples{
\donttest{set.seed(42)
airq.ens <- pre(Ozone ~ ., data = airquality)
get_conditions(airq.ens)}
}
