% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDerivative.R
\name{getDerivative}
\alias{getDerivative}
\title{Get the derivative of a kernel regression function}
\usage{
getDerivative(x, y, bandwidth, newdata = NULL, n = 1, h = NULL)
}
\arguments{
\item{x}{Independent variable.}

\item{y}{Dependent variable.}

\item{bandwidth}{Hyperparameter for the kernel regression.}

\item{newdata}{Optional vector of data points for which we want to calculate the derivative. If \code{NULL}, the derivative will be calculated for \code{x}.}

\item{n}{Order of derivative, should only be between 1 and 8. For \code{n = 0}, function values will be returned.}

\item{h}{Step size. By default, the step size will be set automatically. See details.}
}
\description{
This function numerically approximates the derivative of a function \eqn{f} through the central difference method.
This method computes the average rate of change over a small interval around the point of interest. The function \eqn{f} is estimated via kernel regression given \code{x}, \code{y} and a hyperparameter \code{bandwidth}.
}
\details{
The central difference method as implemented in the \code{pracma} package approximates the derivative of a function \eqn{f} as:

\deqn{\frac{df}{dt} \approx \frac{f(t + h) - f(t - h)}{2h}}

Here, \eqn{h} is a small positive number known as the step size.
The notations \eqn{f(t + h)} and \eqn{f(t - h)} represent the function values at \eqn{t + h} and \eqn{t - h}, respectively.
The smaller \eqn{h} is, the more accurate the approximation should be.
However, if \eqn{h} is too small, then round-off errors from computer arithmetic can
become significant. So, an appropriate balance must be struck.
}
