% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHyperparameters.R
\name{getHyperparameters}
\alias{getHyperparameters}
\title{Get the optimal bandwidth parameters via hyperparameter tuning}
\usage{
getHyperparameters(...)
}
\arguments{
\item{variables}{A character vector with variable names. Hyperparameters will be found for every combination of the specified variables.}

\item{data}{Data in a similar format as the \link{measurements} data.}
}
\description{
This function automatically optimizes the bandwidth for a specified combination of variables.
}
