% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndices.R
\name{getIndices}
\alias{getIndices}
\title{Get the indices related to an index}
\usage{
getIndices(data, i, parameters)
}
\arguments{
\item{data}{A data frame such as \code{\link{measurements}} with the following variables: \code{date}, \code{column}, and \code{depth}.}

\item{i}{The index of the row of interest.}
}
\description{
Get the indices related to an index
}
