% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissing.R
\name{getMissing}
\alias{getMissing}
\title{Interpolate missing measurement data using kernel regression}
\usage{
getMissing(data = PRE::measurements, hyperparameters)
}
\arguments{
\item{data}{The original data we wish to interpolate. Needs to have the same variables as \link{measurements}.}

\item{hyperparameters}{An array of hyperparameters (bandwidths) used in the kernel regression process.}
}
\description{
This function interpolates the missing values of all variable names in \code{hyperparameters}. Also, it returns the derivatives of the smoothing functions used.
}
