% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getN2ON.R
\name{getN2ON}
\alias{getN2ON}
\title{Calculate the volumetric and area N2O-N}
\usage{
getN2ON(data = PRE::measurements, parameters = getParameters())
}
\arguments{
\item{data}{The original data we wish to interpolate. Needs to have the same variables as \link{measurements}.}
}
\description{
This function calculates the volumetric and the area N2O-N from the soil N2O concentration (as stored in the variable \code{N2O} of \code{\link{measurements}}).
}
\details{
In order to calculate the volume N2O-N, the following equation is used:
\deqn{\text N_2 \text {O-N}_\text{volume} =  \frac{28[\text{N}_2\text O]}{\text {R} \cdot \text {T}}}
Where \eqn{\text R} is the gas constant and \eqn{\text T} is the temperature.
}
