% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{Get the PRE parameters.}
\usage{
getParameters(
  BD = 1.686,
  PD = 2.65,
  theta_t = 1 - BD/PD,
  temperature = 298,
  rho = 1260000,
  R = 0.082,
  N2Oatm = 0.2496,
  SPatm = 15.1,
  d18Oatm = 49.6,
  D_fw = 5.07e-06 * exp(-2371/temperature),
  D_fa = 1.436e-05 * (temperature/273.15)^1.81,
  H = (8547000 * exp(-2284/temperature))/(8.3145 * temperature),
  depths = c(7.5, 30, 60, 90, 120),
  SP_nitrification = 34.4,
  SP_denitrification = -2.4,
  eta_SP_diffusion = 1.55,
  eta_SP_reduction = -5.3,
  d18O_nitrification = 36.5,
  d18O_denitrification = 11.1,
  eta_18O_diffusion = -7.79,
  eta_18O_reduction = -16.1
)
}
\arguments{
\item{BD}{Bulk density.}

\item{PD}{Particle density.}

\item{theta_t}{Total soil porosity.}

\item{temperature}{Soil temperature [K].}

\item{rho}{Gas density of N₂O.}

\item{R}{Gas constant [L⋅atm⋅K⁻¹⋅mol⁻¹].}

\item{N2Oatm}{Atmospheric N₂O concentration [ppmv].}

\item{SPatm}{Atmospheric SP concentration.}

\item{d18Oatm}{Atmospheric d18O concentration.}

\item{D_fw}{Diffusivity of N₂O in water.}

\item{D_fa}{Diffusivity of N₂O in air.}

\item{H}{Dimensionless Henry's constant [-].}

\item{depths}{A numeric vector containing the measurement depths [cm].}

\item{SP_nitrification}{Site-preference for nitrification \insertCite{decock2013potential}{PRE}.}

\item{SP_denitrification}{Site-preference for denitrification \insertCite{decock2013potential}{PRE}.}

\item{eta_SP_diffusion}{End member for diffusion of site-preference \insertCite{well2008isotope}{PRE}.}

\item{eta_SP_reduction}{End member for SP reduction \insertCite{denk2017nitrogen}{PRE}.}

\item{d18O_nitrification}{\code{d18O} for nitrification \insertCite{lewicka2017quantifying}{PRE}.}

\item{d18O_denitrification}{\code{d18O} for denitrification \insertCite{lewicka2017quantifying}{PRE}.}

\item{eta_18O_diffusion}{End member for diffusion of \code{d18O} \insertCite{well2008isotope}{PRE}.}

\item{eta_18O_reduction}{End member for \code{d18O} reduction \insertCite{lewicka2017quantifying}{PRE}.}
}
\value{
A list of the PRE data preparation parameters as well as the isotope end members analogous to the function arguments.
}
\description{
This function returns the initial parameters necessary to run the Process Rate Estimator.
To run this function, no function arguments have to be provided, however, if any default values are changed, any dependencies are calculated accordingly.
}
\references{
\insertAllCited{}
}
