% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentNames.R
\name{getPercentNames}
\alias{getPercentNames}
\title{Get Process Names with Quantile Percentages}
\usage{
getPercentNames(
  quantiles,
  process = c("Nitrification", "Denitrification", "Reduction")
)
}
\description{
Get Process Names with Quantile Percentages
}
