% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpe.R
\name{gpe_cv.glmnet}
\alias{gpe_cv.glmnet}
\title{Default penalized trainer for gpe}
\usage{
gpe_cv.glmnet(...)
}
\arguments{
\item{...}{arguments to \code{\link{cv.glmnet}}. \code{x}, \code{y}, \code{weights} and \code{family} will not be used.}
}
\value{
Returns a function with formal arguments \code{x, y, weights, family} and returns a fit object.
}
\description{
Default "penalizer function" generator \code{\link{gpe}} which uses \code{\link{cv.glmnet}}.
}
\seealso{
\code{\link{gpe}}
}
