% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpe.R
\name{gpe_sample}
\alias{gpe_sample}
\title{Sampling Function Generator for gpe}
\usage{
gpe_sample(sampfrac = 0.5)
}
\arguments{
\item{sampfrac}{Fraction of \code{n} to use for sampling. It is the \eqn{\eta / N} in Friedman & Popescu (2008)}
}
\value{
Returns a function that takes an \code{n} argument for the number of observation and an \code{weights} argument for the case weights. The function return a vector of indices
}
\description{
Provides a sample function for \code{\link{gpe}}.
}
\references{
Friedman, J. H., & Popescu, B. E. (2008). Predictive learning via rule ensembles. \emph{The Annals of Applied Statistics}, 916-954.
}
\seealso{
\code{\link{gpe}}
}
