% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpe.R
\name{gpe_tress}
\alias{gpe_tress}
\alias{gpe_linear}
\alias{gpe_earth}
\title{Learner Functions Generators for gpe}
\usage{
gpe_tress(..., remove_duplicates_complements = TRUE, mtry = Inf,
  ntrees = 500, maxdepth = 3L, learnrate = 0.01, parallel = FALSE,
  use_grad = FALSE)

gpe_linear(..., winsfrac = 0.025, normalize = TRUE)

gpe_earth(..., degree = 3, nk = 11, normalize = TRUE, ntrain = 100,
  learnrate = 0.01)
}
\arguments{
\item{...}{Unused}

\item{remove_duplicates_complements}{\code{TRUE} if rules with complementary or duplicate support should be removed from trees}

\item{mtry}{Depth of trees. The argument is passed the tree methods in \code{partykit} package}

\item{ntrees}{Number of trees to fit}

\item{maxdepth}{Maximum dpeth of trees}

\item{learnrate}{Learning rate for methods. It is the \eqn{\nu} parameter in Friedman & Popescu (2008)}

\item{parallel}{\code{TRUE} if basis functions should be found in parallel}

\item{use_grad}{\code{TRUE} if binary outcomes should use gradiant boosting with regression trees when \code{learnrate > 0}. That is, use \code{\link{ctree}} instead of \code{\link{glmtree}} as in Friedman (2000) with a second order Taylor expansion instead of first order as in Chen and Guestrin (2016)}

\item{winsfrac}{Quantiles to winsorize linear terms at. The value should be in \eqn{[0,0.5)}}

\item{normalize}{\code{TRUE} if value should be scaled by \eqn{0.4} times the inverse standard deviation. The \eqn{0.4} is from Friedman & Popescu (2008)}

\item{degree}{Maximum degree of interactions in \code{\link{earth}} model}

\item{nk}{Maximum number of basis functions in \code{\link{earth}} model}

\item{ntrain}{Number of models to fit}
}
\value{
A function that has formal arguments \code{formula, data, weights, sample_func, verbose, family, ...}. The function returns a vector with character where each element is a term for the final formula in the call to \code{\link{cv.glmnet}}
}
\description{
Functions to get "learners" functions for \code{\link{gpe}}.
}
\details{
\code{gpe_tress} provides learners for tree method. Either \code{\link{ctree}} or \code{\link{glmtree}} from the \code{partykit} package will be used.

\code{gpe_linear} provides linear terms for the \code{gpe}.

\code{gpe_earth} provides basis functions where each factor is a hinge function. The model is estimated with \code{\link{earth}}.
}
\references{
Hothorn, T., & Zeileis, A. (2015). partykit: A modular toolkit for recursive partytioning in R. \emph{Journal of Machine Learning Research}, 16, 3905-3909.

Friedman, J. H. (1991). Multivariate adaptive regression splines. \emph{The Annals of Applied Statistics}, 1-67.

Friedman, J. H. (2001). Greedy function approximation: a gradient boosting machine. \emph{The Annals of Applied Statistics}, 1189-1232.

Stanford University. Laboratory for Computational Statistics, & Friedman, J. H. (1993). Fast MARS.

Friedman, J. H., & Popescu, B. E. (2008). Predictive learning via rule ensembles. \emph{The Annals of Applied Statistics}, 916-954.

Chen T., & Guestrin C. (2016). Xgboost: A scalable tree boosting system. \emph{Proceedings of the 22Nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining}. ACM, 2016.
}
\seealso{
\code{\link{gpe}}, \code{\link{rTerm}}, \code{\link{lTerm}}, \code{\link{eTerm}}
}
