% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\docType{data}
\name{hyperparameters}
\alias{hyperparameters}
\title{Hypertuning parameters for the Process Rate Estimator}
\format{
An object of class \code{array} of dimension 5 x 12 x 5.
}
\source{
\url{https://writethesourceurlorthepaper.com}
}
\usage{
data(hyperparameters, package = "PRE")
}
\description{
This is a three-dimensional array, where the first dimension are the depths, the second are the columns, and the third are the specific variables such as \code{N2O}, \code{SP}, or \code{d18O}. These hypertuning parameters (specifically, they are bandwidths for the \link[np:np.regression]{np::npreg} function) were optimized on the \link{measurements} data.
}
\keyword{datasets}
