% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.R
\name{invert}
\alias{invert}
\title{Invert logical vectors}
\usage{
invert(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
This is the equivalent funciton to \code{!}, but it can be called with the pipe operator.
}
