% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{maxdepth.sampler}
\alias{maxdepth.sampler}
\title{Sampling function generator for specifyinf varying maximum tree depth}
\usage{
maxdepth.sampler(av.no.term.nodes = 4L, av.tree.depth = NULL)
}
\arguments{
\item{av.no.term.nodes}{integer of length one.}

\item{av.tree.depth}{integer of length one.}
}
\value{
Returns a random sampling function with single argument 'ntrees',
which can be supplied to the \code{maxdepth} argument of function 
\code{\link{pre}} to specify varying tree depths.
}
\description{
\code{maxdepth.sampler} generates a random sampling function, governed
by a pre-specified average tree depth.
}
\details{
The original RuleFit implementation varying tree sizes for
rule induction. Furthermore, it defined tree size in terms of the number
of terminal nodes. In contrast, function \code{\link{pre}} defines the 
maximum tree size in terms of a (constant) tree depth. Function 
\code{maxdepth.sampler} allows for mimicing the behavior of the
orignal RuleFit implementation. In effect, the maximum tree depth is 
sampled from an exponential distribution with learning rate 
\eqn{\frac{1}{\bar{L}-2}}, where \eqn{(\bar{L}) \geq 2} represents the
average number of terminal nodes for trees in the ensemble. See
Friedman & Popescu (2008, section 3.3).
}
\examples{
## RuleFit default is 4 terminal nodes, on average:
func1 <- maxdepth.sampler()
set.seed(42)
func1(10)
mean(func1(10000))
## Specify 16 terminal nodes on average (equals average maxdepth of 4):
func2 <- maxdepth.sampler(av.no.term.nodes = 16L)
set.seed(42)
func2(10)
mean(func2(10000))
## Specify average maximum tree depth of 3:
func3 <- maxdepth.sampler(av.tree.depth = 3)
set.seed(42)
func3(10)
mean(func3(10000))
}
\references{
Friedman, J. H., & Popescu, B. E. (2008). Predictive learning 
via rule ensembles. \emph{The Annals of Applied Statistics, 2}(3), 916-954.
}
