% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\docType{data}
\name{measurements}
\alias{measurements}
\title{Soil N₂O flux measurements}
\format{
\subsection{\code{measurements}}{

A data frame with 9640 rows and  9 columns:
\describe{
\item{\code{date}}{Date of the measurement [\code{YYYY-MM-DD}].}
\item{\code{column}}{Experimental unit. There are twelve columns in total each with one distinct wheat variety.}
\item{\code{depth}}{Measurement depth [cm]. Either 7.5, 30, 60, 90, or 120 cm.}
\item{\code{increment}}{Height of one conceptual soil layer (borders between the different depth layers).}
\item{\code{variety}}{The wheat variety as a factor. Either \code{"CH Claro"}, \code{"Monte Calme 268"}, \code{"Probus"}, or \code{"Zinal"}.}
\item{\code{moisture}}{The average moisture over a day.}
\item{\code{N2O}}{The corrected N₂O concentration.}
\item{\code{SP}}{Site preference (SP), defined as the difference in isotope value between the central and terminal N atom in the N₂O molecule.}
\item{\code{d18O}}{\eqn{\delta^{18}\text O}, a measure of the deviation in ratio of stable isotopes oxygen-18 (\eqn{^{18}\text O}) and oxygen-16 (\eqn{^{16}\text O}). It is commonly used as an indicator of processes that show isotopic fractionation.}
...
}
}
}
\usage{
data(measurements, package = "PRE")
}
\description{
This data was measured in Zürich for the Process Rate Estimator.
}
\keyword{datasets}
