% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonNegative.R
\name{nonNegative}
\alias{nonNegative}
\title{Apply a function on a vector with removed negative values}
\usage{
nonNegative(x, f = mean)
}
\arguments{
\item{x}{A vector.}

\item{n}{A function to be applied to the vector \code{x}. By default, this is \link[base:mean]{base::mean}.}
}
\description{
This function applies a function \code{f} to a vector \code{x}, but first sets all negative values in \code{x} to zero.
}
