% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.longPRE.R
\name{plot.longPRE}
\alias{plot.longPRE}
\title{Generic plot function for a \code{longPRE} class object}
\usage{
\method{plot}{longPRE}(x, which = c("Nitrification", "Denitrification", "Reduction"))
}
\arguments{
\item{x}{The output of the process rate estimator (over time) function \code{longPRE}.}
}
\description{
Generic plot function for a \code{longPRE} class object
}
