% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpe_miscs.R
\name{predict.gpe}
\alias{predict.gpe}
\title{Predicted Values Based on gpe Ensemble}
\usage{
\method{predict}{gpe}(object, newdata = NULL, type = "link",
  penalty.par.val = "lambda.1se", ...)
}
\arguments{
\item{object}{of class \code{\link{gpe}}}

\item{newdata}{optional new data to compute predictions for}

\item{type}{argument passed to \code{\link{predict.cv.glmnet}}}

\item{penalty.par.val}{argument passed to \code{s} argument of \code{\link{predict.cv.glmnet}}}

\item{...}{Unused}
}
\description{
Predict function for \code{\link{gpe}}
}
\details{
The initial training data is used if \code{newdata = NULL}.
}
\seealso{
\code{\link{gpe}}
}
