% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.longPRE.R
\name{print.longPRE}
\alias{print.longPRE}
\title{Generic print function for a \code{longPRE} class object}
\usage{
\method{print}{longPRE}(x, replaceNegative = TRUE)
}
\arguments{
\item{x}{The output of the process rate estimator function \code{longPRE}.}

\item{replaceNegative}{Logical (\code{TRUE}/\code{FALSE}). Should the negative values be replaced by zero in order to calculate the mean rates?}
}
\description{
Generic print function for a \code{longPRE} class object
}
