% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressbar.R
\name{progressbar}
\alias{progressbar}
\title{Show the progress in the console}
\usage{
progressbar(i, n)
}
\arguments{
\item{i}{The current index.}

\item{n}{The maximum index.}
}
\description{
Show the progress in the console
}
