% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateEquations.R
\name{stateEquations}
\alias{stateEquations}
\title{State equation set}
\usage{
stateEquations(
  x = c(N2Onit = NA, N2Oden = NA, N2Ored = NA),
  e = unlist(getEpsilons()),
  fluxes = list(F_top_in = NA, F_bottom_in = NA, F_out = NA, SP_top = NA, SP = NA,
    SP_bottom = NA, N2ONarea = NA, d18O_top = NA, d18O_bottom = NA, d18O = NA)
)
}
\arguments{
\item{x}{Vector with starting values of \code{N2Onit}, \code{N2Oden}, and \code{N2Ored}.}

\item{e}{Vector of the isotope end members.}

\item{fluxes}{Values of \code{N2ONarea}, \code{SP} and \code{d18O} as well as their fluxes from other increments.}

\item{derivatives}{Derivatives of \code{N2ONarea}, \code{SP} and \code{d18O} with respect to time.}
}
\description{
The state equation set are the core of the process rate estimator (PRE).
}
\details{
The state equation set incorporates three distinct equations.
\deqn{\frac{\Delta[N_2O]}{\Delta t}}
}
