% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateFunctions.R
\name{stateFunctions}
\alias{stateFunctions}
\title{State function set}
\usage{
stateFunctions(
  x,
  e = getEpsilons(),
  Ftopin,
  Fbottomin,
  Fout,
  dN2O_dt,
  SPtopin,
  SP,
  SPbottomin,
  N2O,
  dSP_dt,
  d18Otopin,
  d18Obottomin,
  d18O,
  d18O_dt
)
}
\description{
State function set
}
