% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFluxes.R
\name{calculateFluxes}
\alias{calculateFluxes}
\title{Calculate fluxes from measurement data}
\usage{
calculateFluxes(data = PRE::measurements, parameters = PRE::getParameters())
}
\arguments{
\item{data}{Measurement data from which one wants to calculate fluxes.
Needs to include variables \code{date}, \code{column}, \code{depth}, \code{increment}, \code{moisture}, \code{N2O}, \code{CO2}, \code{SP}, and \code{d18O}. For further details, read the description of the \code{\link{measurements}} data.}

\item{parameters}{A set of parameters as returned by the function \code{\link{getParameters}}. These will be used in calculating the returned flux data.}
}
\description{
This function calculates fluxes from measurement data (e.g. \code{PRE::measurements}).
}
