% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_pre_model.R
\docType{data}
\name{caret_pre_model}
\alias{caret_pre_model}
\title{Model set up for train function of package caret}
\format{An object of class \code{list} of length 17.}
\usage{
caret_pre_model
}
\description{
\code{caret_pre_model} provides a model setup for the train function of
package caret
}
\examples{
\dontrun{
 
 library("caret")
 ## Generate some random data:
 set.seed(42)
 x <- data.frame(x1 = rnorm(50), x2 = rnorm(50), noise = rnorm(50))
 y <- x$x1 + x$x2 + rnorm(50)
 
 ## Fit caret using only default settings:
 set.seed(42)
 prefit1 <- train(x = x, y = y, method = caret_pre_model, 
                  trControl = trainControl(number = 1))
 prefit1                 
 
 ## Fit caret using customized tuneGrid:
 set.seed(42)
 tuneGrid <- caret_pre_model$grid(x = x, y = y, 
                  type = c("linear", "rules", "both"),
                  maxdepth = 2:5)
 tuneGrid
 prefit2 <- train(x = x, y = y, method = caret_pre_model, 
                  trControl = trainControl(number = 1),
                  tuneGrid = tuneGrid)
 prefit2
}
}
\keyword{datasets}
