% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEpsilons.R
\name{getEpsilons}
\alias{getEpsilons}
\title{Get the isotope end members}
\usage{
getEpsilons(
  SPnit = 34.4,
  SPden = -2.4,
  eSPdiffusion = 1.55,
  eSPred = -5.3,
  d18Onit = 36.5,
  d18Oden = 11.1,
  e18Odiffusion = -7.79,
  e18Ored = 3 * eSPred
)
}
\arguments{
\item{SPnit}{Site preference nitrification \insertCite{lewicka2017quantifying}{PRE}.}

\item{SPden}{Lewicka-Szceback for SPden.}

\item{eSPdiffusion}{Well et al. 2008 for eSPdiffusion.}

\item{eSPred}{Lewicka-Szceback for eSPred.}

\item{d18Onit}{Lewicka-Szceback for d18Onit.}

\item{d18Oden}{Lewicka-Szceback for d18Oden.}

\item{e18Odiffusion}{e18Odiffusion \insertCite{well2008isotope}{PRE}.}

\item{e18Ored}{Lewicka-Szceback e18Ored.}
}
\value{
A list of the eight isotope end members analogous to the function arguments.
}
\description{
Get the isotope end members
}
\references{
\insertAllCited{}
}
