% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMissing.R
\name{getMissing}
\alias{getMissing}
\title{Interpolate missing measurement data using kernel regression}
\usage{
getMissing(data = getN2ON(), hyperparameters = PRE::hyperparameters)
}
\arguments{
\item{data}{The original data we wish to interpolate. Needs to be a data frame that is returned by the \link{getN2ON} function.}

\item{hyperparameters}{A three-dimensional array of hyperparameters (bandwidths) used in the kernel regression process. For references, see \link{hyperparameters}.}
}
\description{
This function interpolates the missing values of all variable names in \code{hyperparameters}.
Also, it returns the derivatives of the smoothing functions used.
For both procedures, it calls the \link{getDerivative} function -- simply with \code{n = 0} and \code{n = 1}.
}
