% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{getParameters}
\alias{getParameters}
\title{Get the PRE parameters.}
\usage{
getParameters(
  BD = 1.686,
  PD = 2.65,
  theta_t = 1 - BD/PD,
  temperature = 298,
  rho = 1260000,
  R = 0.082,
  N2Oatm = 0.2496,
  SPatm = 15.1,
  d18Oatm = 49.6,
  D_fw = 5.07e-06 * exp(-2371/temperature),
  D_fa = 1.436e-05 * (temperature/273.15)^1.81,
  H = (8547000 * exp(-2284/temperature))/(8.3145 * temperature),
  depths = c(7.5, 30, 60, 90, 120)
)
}
\arguments{
\item{BD}{Bulk density.}

\item{PD}{Particle density.}

\item{theta_t}{Total soil porosity.}

\item{temperature}{Soil temperature [K].}

\item{rho}{Gas density of N₂O.}

\item{R}{Gas constant [L⋅atm⋅K⁻¹⋅mol⁻¹].}

\item{N2Oatm}{Atmospheric N₂O concentration [ppmv].}

\item{SPatm}{Atmospheric SP concentration.}

\item{d18Oatm}{Atmospheric d18O concentration.}

\item{D_fw}{Diffusivity of N₂O in water.}

\item{D_fa}{Diffusivity of N₂O in air.}

\item{H}{Dimensionless Henry's constant [-].}

\item{depths}{A numeric vector containing the measurement depths [cm].}
}
\description{
This function returns the initial parameters necessary to run the Process Rate Estimator.
No function arguments have to be provided, however, if any default values are changed, the set of PRE parameters are calculated accordingly.

The function returns a list of parameters that are used to run the Process Rate Estimator (PRE).
The function return values are the same as the function arguments, but computed dependently.
}
