% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPercentNames.R
\name{getPercentNames}
\alias{getPercentNames}
\title{Get Process Names with quantile percentages}
\usage{
getPercentNames(
  quantiles,
  process = c("Nitrification", "Denitrification", "Reduction")
)
}
\description{
This is an internal function which takes the quantiles as an argument and creates variable names of the quantiles and processes combinations.
}
