% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\docType{data}
\name{hyperparameters}
\alias{hyperparameters}
\title{Hypertuning parameters for the Process Rate Estimator}
\format{
An object of class \code{array} of dimension 5 x 12 x 5.
}
\source{
\url{https://writethesourceurlorthepaper.com}
}
\usage{
data(hyperparameters, package = "PRE")
}
\description{
These hypertuning parameters (hyperparameters) were optimized on the \link{measurements} data.
}
\keyword{datasets}
