% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{interact}
\alias{interact}
\title{Calculate interaction statistics for user-specified variables}
\usage{
interact(object, varnames = NULL, nullmods = NULL,
  penalty.par.val = "lambda.1se", quantprobs = c(0.05, 0.95), plot = TRUE,
  col = c("yellow", "blue"), ylab = "Interaction strength",
  main = "Interaction test statistics", parallel = FALSE, k = 10,
  verbose = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{\link{pre}}.}

\item{varnames}{character vector. Names of variables for which interaction
statistics should be calculated. If \code{NULL}, interaction statistics for
all predictor variables with non-zeor coefficients will be calculated (which
may take a long time).}

\item{nullmods}{object with bootstrapped null interaction models, resulting
from application of \code{bsnullinteract}.}

\item{penalty.par.val}{character. Which value of the penalty parameter
criterion should be used? The value yielding minimum cv error
(\code{"lambda.min"}) or penalty parameter yielding error within 1 standard
error of minimum cv error ("\code{lambda.1se}")? Alternatively, a numeric 
value may be specified, corresponding to one of the values of lambda in the 
sequence used by glmnet, for which estimated cv error can be inspected by 
running \code{object$glmnet.fit} and \code{plot(object$glmnet.fit)}.}

\item{quantprobs}{numeric vector of length two. Probabilities that should be
used for plotting the range of bootstrapped null interaction model statistics.
Only used when \code{nullmods} argument is specified and \code{plot = TRUE}.
The default yields sample quantiles corresponding to .05 and .95 probabilities.}

\item{plot}{logical. Should interaction statistics be plotted?}

\item{col}{character vector of length one or two. Color for plotting 
interaction statistics. The first color specified is used to plot the 
interaction statistic from the training data, the second color specifed
is used to plot the interaction statistic distribution from the bootstrapped
null interaction models. Only used when \code{plot = TRUE}. Only the first 
element of vector is used if \code{nullmods = NULL}.}

\item{ylab}{character string. Label to be used for plotting y-axis.}

\item{main}{character. Main title for the bar plot.}

\item{parallel}{logical. Should parallel foreach be used? Must register
parallel beforehand, such as doMC or others.}

\item{k}{integer. Calculating interaction test statistics is a computationally
intensive, so  calculations are split up in several parts to prevent memory
allocation errors. If a memory allocation error still occurs, increase k.}

\item{verbose}{logical. Should progress information be printed to the
command line?}

\item{...}{Additional arguments to be passed to \code{barplot}.}
}
\value{
If nullmods is not specified, the function returns the interaction
test statistic. If nullmods is specified, the function returns a list,
with elements \code{$H}, which is the test statistic of the interaction
strength, and \code{$nullH}, which is a vector of test statistics of the
interaction in each of the bootstrapped null interaction models. 
If \code{plot = TRUE} (the default), a barplot is created, with the 
interaction test statistic based on the selected ensemble and, if 
\code{nullmods} is specified, the distribution of interaction test statistics 
of the bootstrapped null interaction models, within the quantiles specified. 
This allows for testing the null hypothesis of no interaction effect for 
each of the input variables. If for an input variable the interaction test 
statistic based on the selected ensemble exceeds the upper limit of the error 
bar plotted for the bootstrapped null interaction models, the null hypothesis
of no interaction with other input variables can be rejected at an alpha 
level of quantprobs[2].
}
\description{
\code{interact} calculates test statistics for assessing the strength of
interactions between the input variable(s) specified, and all other input
variables.
}
\details{
Can be computationally intensive, especially when nullmods is specified,
in which case setting \verb{parallel = TRUE} may improve speed.
}
\examples{
\donttest{
 set.seed(42)
 airq.ens <- pre(Ozone ~ ., data=airquality[complete.cases(airquality),])
 interact(airq.ens, c("Temp", "Wind", "Solar.R"))}
}
\seealso{
\code{\link{pre}}, \code{\link{bsnullinteract}}
}
