% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longPRE.R
\name{longPRE}
\alias{longPRE}
\title{Run the Process Rate Estimator on all dates for one experimental unit}
\usage{
longPRE(data, column, depth, n = 100, nonNegative = FALSE)
}
\arguments{
\item{data}{A data frame with all relevant variables to run the Process Rate Estimator.}

\item{column}{The column name for which we want to run the PRE.}

\item{depth}{The depth name for which we want to run the PRE.}

\item{n}{The number of samples solved with the \link[BB:multiStart]{BB::multiStart} solver.}

\item{nonNegative}{Should negative solutions be removed before retrieving the quantiles?}

\item{date}{The date (\code{DD-MM-YYYY}) for which we want to run the PRE.}

\item{quantiles}{The probabilities of the quantiles to be returned from the sampled set.}

\item{verbose}{Should a success message be printed after a model run?}
}
\description{
This function runs the Process Rate Estimator (PRE) on a specific depth and column, but on all available dates.
}
