% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurements.R
\docType{data}
\name{measurements}
\alias{measurements}
\title{Soil N₂O flux measurements}
\format{
\subsection{\code{measurements}}{

A data frame with 9558 rows and  14 columns:
\describe{
\item{\code{date}}{Date of the measurement [\code{YYYY-MM-DD}].}
\item{\code{column}}{Experimental unit.}
\item{\code{depth}}{Measurement depth [cm]. Either 7.5, 30, 60, 90, or 120 cm.}
\item{\code{increment}}{Height of one conceptual soil layer (borders between the different depth layers).}
\item{\code{variety}}{The wheat variety as a factor. Either \code{"CH Claro"}, \code{"Monte Calme 268"}, \code{"Probus"}, or \code{"Zinal"}.}
\item{\code{moisture}}{The average moisture over a day.}
\item{\code{N2O}}{The corrected N₂O concentration.}
\item{\code{CO2}}{The corrected CO₂ concentration.}
\item{\code{SP}}{The site preference.}
\item{\code{d18O}}{The δ¹⁸O concentration.}
\item{\code{d15Nbulk}}{The bulk δ¹⁵N concentration.}
\item{\code{d15Nalpha}}{The δ¹⁵N α concentration.}
\item{\code{d15Nbeta}}{The δ¹⁵N β concentration.}
...
}
}
}
\source{
\url{https://writethesourceurlorthepaper.com}
}
\usage{
data(measurements, package = "PRE")
}
\description{
This data was measured in Zürich for the Process Rate Estimator.
}
\keyword{datasets}
