% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PRE.R
\name{plot.PRE}
\alias{plot.PRE}
\title{Generic plot function for a \code{PRE} class object}
\usage{
\method{plot}{PRE}(x, col = "grey")
}
\arguments{
\item{x}{The output of the process rate estimator function \code{PRE}.}
}
\description{
Generic plot function for a \code{PRE} class object
}
