% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PRE.R
\name{print.PRE}
\alias{print.PRE}
\title{Generic print function for a \code{PRE} class object}
\usage{
\method{print}{PRE}(x)
}
\arguments{
\item{x}{The output of the process rate estimator function \code{PRE}.}
}
\description{
Generic print function for a \code{PRE} class object
}
