% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stateEquations.R
\name{stateEquations}
\alias{stateEquations}
\title{State equation set}
\usage{
stateEquations(
  x = c(Nitrification = NA, Denitrification = NA, Reduction = NA),
  parameters = getEpsilons(),
  fluxes = list(F_top_in = NA, F_bottom_in = NA, F_out = NA, SP_top = NA, SP = NA,
    SP_bottom = NA, N2ONarea = NA, d18O_top = NA, d18O_bottom = NA, d18O = NA,
    N2ONarea_derivative = NA, SP_derivative = NA, d18O_derivative = NA)
)
}
\arguments{
\item{x}{A vector with the named values of N₂O \code{Nitification}, \code{Denitrification}, and \code{Reduction}.}

\item{parameters}{A list of parameters containing the isotope end members. By default, the \code{\link{getParameters}} function is used to return these values.}

\item{fluxes}{A list of \code{N2ONarea}, \code{SP} and \code{d18O} as well as their fluxes from other increments and derivative. Essentially, this is one row of the data as it's returned by the \code{\link{calculateFluxes}} function as a list.}
}
\description{
The state equation set are the core of the process rate estimator (PRE).
}
\details{
The state equation set incorporates three distinct equations.
\deqn{\frac{\Delta[N_2O]}{\Delta t}}
}
